Create README.md
This project addresses the challenges of using AI to recognize human activity in dementia care due to a lack of trust in decisions made by AI. The models are able to recognise both normal and abnormal behaviors related to dementia. With the use of explainable machine learning algorithms, the decisions of these models were explained and interpreted.